clean up rails_helper, refactored omniauthstub method into spec helper, and configure rails_helper to include vcr config
